backend/requirements.txt
APScheduler==3.11.0
fastapi
fpdf==1.7.2
httpx==0.28.1
langchain==1.2.7
langchain_anthropic==1.3.1
langchain_core==1.2.7
langchain_google_genai==4.2.0
numpy==2.4.1
pandas==3.0.0
protobuf==6.33.4
pydantic==2.12.5
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
quickchart-io
scipy==1.17.0
starlette
websockets==15.0.1
streamlit
plotly
